% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-tools.R
\name{opcode_table}
\alias{opcode_table}
\title{Obtain the OP codes of compiled bytecode}
\usage{
opcode_table(bcode)
}
\arguments{
\item{bcode}{bytecode or something that can be compiled to byte code}
}
\description{
Obtain the OP codes of compiled bytecode
}
