% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-tools.R
\name{tidy_bytecode}
\alias{tidy_bytecode}
\title{Obtain the OP codes of compiled bytecode}
\usage{
tidy_bytecode(bcode)
}
\arguments{
\item{bcode}{bytecode or something that can be compiled to byte code}
}
\description{
Obtain the OP codes of compiled bytecode
}
